Move Fft trait
